+++ src/chroniclers/backends/custom_backend.nim
@@ -1,0 +1,6 @@
+## Replace this module with a custom logging backend using NimScript's patchFile.
+
+{.
+  error:
+    "The chroniclers custom backend is empty. Use -d:chroniclers.logBackendCustom and add patchFile(\"chroniclers\", \"custom_backend\", \"path/to/backend\") to your application's config.nims."
+.}